Fixes related to the saturation flag
* Adjust the saturation value according to the flux scale
* Use the original image to set the saturation flag
* Compare with >=

If the flux scale happens to be <= 1, then the saturated pixels will
always have a value <= to the saturation. Using the background
substracted image will move also the value below the limit.

With flux scales > 1 happened to work fine as the saturated pixels
would be well above the saturation threshold.