Do an actual convolution, not a correlation
The confusion arised because OpenCV filter2D claims that

> Convolves an image with the kernel.

But in fact

> The function does actually compute correlation, not the convolution

As that was the original implementation, the error propagated to the
other implementations.

As the OpenCV documentation hints, the solution is to flip (rotate 180)
the kernel.

For the DFT, however, as it internally mirrors before doing the FFT, we
just drop the existing mirroring, as mirror(mirror(image)) = image.