Parameters that are not fitted set to NaN
For instance, a source may not be in any frame for a given band. The
code will, correctly, *not* fit any model specific to that band.
However, the output parameter value will be the initial guess (i.e.
isophotal flux on the detection image), which can be confusing.

With this change, these parameters are "reaped" and set to NaN. If there
are any dependent parameters, the NaN will propagate.