Multithreaded property computation for grouping/cleaning (#290)
* Demangle property names

* Use Alexandria ThreadPool for measurements

First step to share the thread pool with the deblending

* Put prefetcher in place

* Prefetcher must control ProcessSourcesEvent

To make sure everything before has been done

* Change unsigned to int64. Add docs.

* Move prefetcher to SEImplementation

* Fix build on co7

* Enforce ordering on prefetcher

* Log when levmar can not be used multithreaded

* Let workers go on while the message is sent

* Use object address for the prefetcher queues

* group_id counter should be atomic

* So should be source_id

* It is better to release lock before notifying

Avoid wake+lock

* Make ProgressCounter thread safe

* Fix build on CentOS7

* Remove leftover check

It was originally intended for an end-of-image signaled by a null source,
but that was obsoleted. Since m_done wasn't explicitly initialized, it just
happened to be true by chance on several platforms except centos8.